JCS-14035 - Fail to get password expiry date when using connect string (#225)
Bug - Fail to get password expiry date for OPSS user when using connect
string

Note that since the DB service name is not guaranteed to include the PDB
name ( I proved this by using a connect string w/o the PDB name in it to
successfully create a WLS for OCI instance). Therefore, the PDB name
must be asked for.

Also note that the validation change added will not be executed, but to
limit the scope of the changes I updated the validation only and didn't
try to also add in the validator. I suspect that the validation was
never added in order to ensure that 11g databases, which don't have a
PDB can be allowed.

Tested that when setting a connect string the error occurred. After the
fix, with PDB name provided, the error did not occur.