Add new flag for autoscaling alarm state defaulted to true (#277)
The monitoring alarm is being created as enabled only when the option
"Create OCI policies" checkbox is selected. The problem seems to be in
monitoring_alarms.tf line 11 because the flag "is_enabled" is associated
directly to var.create_policies which doesn't make sense. This PR is to
add a new variable independent to var.create_policies so when customer
clear the "Create OCI policies" checkbox the alarms created for
autoscaling remain enabled.

This option is intentionally hidden to the customer as we don't have a
use case for customers to create autoscaling with disabled alarms and
these are defaulted to true. The variables are configured in case of
advanced customers who may want to customize the stack can easily
override those variables.

Testing with OCI Checkbox enabled:

-     Verify new variable is not shown in ORM UI.
**12.2.1.4**

![345111004-d5f0e922-7f4f-4300-8d04-ca16f03dd73c](https://github.com/oracle-quickstart/oci-weblogic-server/assets/44181317/490e30e1-a844-4ad8-9418-a98f7f7d16e5)

**14.1.1**
<img width="772" alt="image"
src="https://github.com/oracle-quickstart/oci-weblogic-server/assets/106961326/f14cf051-00c5-4f8b-a06c-c3493b129e3d">

Usually undefined variables are shown at the bottom, this image is
evidence that new variable is hidden from customer and default value
(true) is used.

- Verify that provisioning succeed with OCI policy checkbox enabled and
when autosacling is enabled alarms are enabled after provisioning

![345111695-9cd6395d-ccc1-439e-8ab7-31f498a31e0d](https://github.com/oracle-quickstart/oci-weblogic-server/assets/44181317/d43eef16-d2cf-438b-979b-30b0841d8f86)

![345111870-864383e0-e987-416b-a2c0-44251b024c1f](https://github.com/oracle-quickstart/oci-weblogic-server/assets/44181317/779cc653-aa9a-452a-9d04-38199d29730b)

- Verify that provisioning succeed with OCI policy checkbox disabled and
when autosacling is enabled alarms are enabled after provisioning

![345112076-4af137ca-707c-4f82-b977-7d73c0389cd9-1](https://github.com/oracle-quickstart/oci-weblogic-server/assets/44181317/a97ff757-5af2-4453-b2e0-56e4fbcaed34)

![345112160-0c45d6ef-8037-49b9-8cd0-6886ac7dd9f5](https://github.com/oracle-quickstart/oci-weblogic-server/assets/44181317/b9915eb2-979f-4cfb-a4a3-4654f52ad5f3)

---------

Co-authored-by: Mahuwa-Barman <mahuwa.barman@oracle.com>