JCS-14582 - Manage instances in dynamic group policies (#286)
JCS-14582 - WLS for OCI dynamic group policy for instances does not
allow volumes to be added to the instance. This causes issues for
cloning scripts when volumes need to be added and removed from the
instances using the OCI API. Verified by creating a WLS for OCI instance
with "OCI Polcies" checkbox enabled and cloning that instance.