Modify the script to add additional support for network validation (#347)
Following are the list of changes that are being done - 
a.) Added a check to see if the subnet is created with the option "Use
DNS hostnames in this VCN" enabled
b.) Made modification to the code to correctly verify the option for
CustomDNSResolver
c.) When the LPG's are peered, they are in the status as "Peered". MAde
the code change to check if the status is "Peered" or not.
d.) Ensured that the Route table associated with the subnet of WLS and
DB does have a Route rule to the LPG with the corresponding CIDR. Also,
it will verify if the CIDR of the Route rule does fall with the VCN's
range.
e.) Also added a check to ensure that the WLS Subnet SecList/NSG does
have an egress rule to the DB Port and DB CIDR.
f.) Added a new option to specify the WLS LPG OCID and also added a
check to make sure that both the LPG OCID's are provided in the command.