Removing the option of APM in edit stack (#349)
Currently when editing a stack the option to "Enable Application
Performance Monitoring" is presented. The decision is made to hide the
option while editing the stack. In this MR, we will be making changes to
the terraform to present the option of APM only during the creation of
stack.

Testing - 
a.) Created the stack with APM enabled. Verified that the APM agent is
added on all the servers and its being monitored by APM from dashboard.
b.) Scaled up the nodes and verified that the APM agent was installed on
the newly added server and that is being shown in the APM dashboard for
monitoring.
c.) Verified from the UI that the checkbox of APM is being disabled in
the edit stack screen