Topic ocir user edit (#350)
When the OCIR user is specified for a BOAT user (e.g.
bmc_operator_access/myuser) for autoscaling the tenancy namespace for
the tenancy on which a WLS for OCI stack is applied is added to the user
name (e.g. paasprodjcs/bmc_operator_access/myuser). Since the
bmc_operator_access is the tenancy namespace the addition of the
additional tenancy namespace causes OCIR requests to fail.

With agreement from the proposal suggested by Adrian, we are going with
the following change -
If the ocir_username entered contains '/', then we assume that the
customer has entered the complete info including the tenancy_namespace,
identity domain and the username.
If the ocir_username doesn't have a '/', then we append the namespace to
it.

Testing - 
a.) Created a stack with the tenancy namespace added. - Podman login was
successful and the namespace was not appended again.
b.) Created a stack without the namespace in the ocir_user variable -
Stack provisioning did append the namespace and in our case it was the
tenancy namespace where the stack is created